3_commit

+++ src/main/java/com/pocket_plant/backend/dto/MemberTokenResponse.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 public class MemberTokenResponse {
 
     private String serviceToken; // 우리 서비스가 발행한 인증 토큰 (예: JWT)
-    private int userId; // 로그인된 회원의 고유 ID
+    private Long userId; // 로그인된 회원의 고유 ID
     private String nickname; // 로그인된 회원의 닉네임
     private String email; // 로그인된 회원의 이메일
     private String profileImageUrl; // 로그인된 회원의 프로필 이미지 URL
